--- 17.React/desafio2/desafio2/src/components/Alert.jsx
@@ -1,1 +1,0 @@
-
