--- 17.React/desafio2/desafio2/src/components/Registro.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-import SocialButton from "./components/SocialButton";
-import Formulario from "./components/Formulario";
-
-const Registro = () => {
-  return (
-    <div>
-      <SocialButton />
-      <Formulario />
-    </div>
-  );
-};
-
-export default Registro;
