--- 5git/clase/index.html
@@ -1,25 +1,0 @@
-Github es una red social 
-    Exponer nuestro trabajo al mundo 
-    Trabajar de forma colaborativa 
-
-Desafío: publicar código 
-
-Prompt: una terminal donde se puede preguntar algo y obtener respuesta
-    El power shell te deja en tu directorio de usuario 
-    Comandos:
-        ls - listar qué tengo en el contenido del directorio de ususario
-        pwd - para saber dónde estás 
-        cd.. - (change directory) subir un nivel en el directorio, para llegar a C hay que hacerlo 2 veces
-        cd fullstack_js - para ir al directorio
-
-        cd \
-        cd fullstack_js\5git\clase
-
-        mkdir - make directory, para crear una carpeta y luego ponemos el nombre (ej mkdir img)
-        rm - eliminar los directorios 
-        echo- Crear un nuevo archivo vacío 
-        clear - limpiar pantalla
-        
-        git init - cuando se ponga verde es porque ya estamos con git en vs
-        git add index.html
-        git commit -m "nombre del proyecto" - empaquetar la versión
