+++ desafio2/src/App.jsx
@@ -1,13 +1,9 @@
 import { useState } from "react";
-import reactLogo from "./assets/react.svg";
-import viteLogo from "/vite.svg";
 import "./App.css";
 import "bootstrap/dist/css/bootstrap.min.css";
-import Registro from "./components/Registro";
+import { Registro } from "./components/registro/Registro";
 
 function App() {
-  const [count, setCount] = useState(0);
-
   return (
     <>
       <Registro></Registro>
