+++ desafio2/src/components/alert/Alert.jsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Alert = () => {
+  return <div>Alert</div>;
+};
+
+export default Alert;
