+++ desafio2/src/components/icons/Icons.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const Icons = () => {
+  return <div>Icons</div>;
+};
