+++ desafio2/src/components/registro/Registro.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { SocialButton } from "../socialbutton/SocialButton";
+import { Formulario } from "../formulario/Formulario";
+
+export const Registro = () => {
+  return (
+    <div>
+      <SocialButton />
+      <Formulario />
+    </div>
+  );
+};
