+++ desafio2/src/components/socialbutton/SocialButton.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const SocialButton = () => {
+  return <div>SocialButton</div>;
+};
